requirements.txt
boto3
python-daemon
